move certificate classes to their own namespace
